Version 1.0 - All Basic Features Implemented
Add/Remove/Modify/Refresh all work. Basic error checking/input validation is implemented. A story list was added along with a basic story browsing form containing a link to the full website. Features I'd like to add: custom Rss Icon on main page window. Pulling the icon from the rss feed, storing it in the db, and displaying it on the single rss form icon and page, as well as the icon on the story browsing form.