1.3: Refactor functions & minor changes to scroll.
In this revision, I made minor changes to scrolling in order to better capture focus. I also tried to separate functions/subs that are performing actions from methods that are handling events. This was to prevent event handlers from conflicting when trying to run functions/subs that handle different kinds of events than the parent function/sub. By making functions to perform the actions, the sender and eventargs no longer need to be passed and the issue is resolved.